IGNITE-23593 Use CacheQuery instead of GridCacheQueryAdapter (#11639)
